requirements.txt
Flask
python-dotenv
firebase-admin
gunicorn